refactor: behavioral and performance improvements
