+++ .gitignore
@@ -19,3 +19,6 @@ Thumbs.db
 
 # Tauri
 /gen/
+
+# Examples
+examples/**/Cargo.lock
