+++ android/example/build.gradle.kts
@@ -5,13 +5,13 @@ plugins {
 }
 
 android {
-    namespace = "com.velocitysystems.downloadmanager.example"
-    compileSdk = 36
+    namespace = "org.silvermine.downloadmanager.example"
+    compileSdk = 35
 
     defaultConfig {
-        applicationId = "com.velocitysystems.downloadmanager.example"
-        minSdk = 26
-        targetSdk = 36
+        applicationId = "org.silvermine.downloadmanager.example"
+        minSdk = 24
+        targetSdk = 35
         versionCode = 1
         versionName = "1.0"
     }
@@ -42,10 +42,9 @@ android {
 
 dependencies {
     implementation(project(":lib"))
-    implementation("androidx.core:core-ktx:1.15.0")
     implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.8.7")
     implementation("androidx.lifecycle:lifecycle-viewmodel-compose:2.8.7")
-    implementation("androidx.activity:activity-compose:1.9.3")
+    implementation("androidx.activity:activity-compose:1.9.2")
     implementation(platform("androidx.compose:compose-bom:2024.12.01"))
     implementation("androidx.compose.ui:ui")
     implementation("androidx.compose.ui:ui-graphics")
