+++ android/example/src/main/AndroidManifest.xml
@@ -22,8 +22,5 @@
             </intent-filter>
         </activity>
 
-        <service
-            android:name="androidx.work.impl.foreground.SystemForegroundService"
-            android:foregroundServiceType="dataSync" />
     </application>
 </manifest>
