--- android/example/src/main/java/com/velocitysystems/downloadmanager/example/MainActivity.kt
@@ -1,19 +1,0 @@
-package com.velocitysystems.downloadmanager.example
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.material3.MaterialTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            MaterialTheme {
-                DownloadsScreen()
-            }
-        }
-    }
-}
