+++ android/example/src/main/java/org/silvermine/downloadmanager/example/MainActivity.kt
@@ -1,0 +1,19 @@
+package org.silvermine.downloadmanager.example
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.material3.MaterialTheme
+
+class MainActivity : ComponentActivity() {
+   override fun onCreate(savedInstanceState: Bundle?) {
+      super.onCreate(savedInstanceState)
+      enableEdgeToEdge()
+      setContent {
+         MaterialTheme {
+            DownloadsScreen()
+         }
+      }
+   }
+}
