--- android/lib/src/main/java/com/velocitysystems/downloadmanager/DownloadException.kt
@@ -1,9 +1,0 @@
-package com.velocitysystems.downloadmanager
-
-/**
- * Exceptions thrown by download operations.
- */
-sealed class DownloadException(message: String) : Exception(message) {
-    /// Download item was not found for the given path.
-    class NotFound(path: String) : DownloadException("Download not found: $path")
-}
