+++ android/lib/src/main/java/org/silvermine/downloadmanager/DownloadStore.kt
@@ -1,0 +1,84 @@
+package org.silvermine.downloadmanager
+
+import android.content.Context
+import android.util.AtomicFile
+import android.util.Log
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
+import java.io.File
+
+/**
+ * Thread-safe store for download items backed by an atomic JSON file.
+ *
+ * All public methods are synchronized to ensure consistency when accessed
+ * from multiple threads (e.g. WorkManager workers and the main thread).
+ * Mirrors the iOS DownloadStore actor pattern.
+ */
+internal class DownloadStore(context: Context) {
+   private val json = Json { ignoreUnknownKeys = true }
+   private val file = AtomicFile(File(context.filesDir, STORE_FILENAME))
+   private val downloads = mutableMapOf<String, DownloadItem>()
+
+   init {
+      load()
+   }
+
+   @Synchronized
+   fun list(): List<DownloadItem> = downloads.values.toList()
+
+   @Synchronized
+   fun findByPath(path: String): DownloadItem? = downloads[path]
+
+   @Synchronized
+   fun append(item: DownloadItem) {
+      downloads[item.path] = item
+      save()
+   }
+
+   @Synchronized
+   fun update(item: DownloadItem, persist: Boolean = true) {
+      if (downloads.containsKey(item.path)) {
+         downloads[item.path] = item
+      }
+      if (persist) {
+         save()
+      }
+   }
+
+   @Synchronized
+   fun remove(item: DownloadItem) {
+      downloads.remove(item.path)
+      save()
+   }
+
+   private fun load() {
+      try {
+         val bytes = file.readFully()
+         val items: List<DownloadItem> = json.decodeFromString(String(bytes))
+         downloads.clear()
+         for (item in items) {
+            downloads[item.path] = item
+         }
+      } catch (e: Exception) {
+         Log.e(TAG, "Failed to load download store: ${e.message}")
+      }
+   }
+
+   private fun save() {
+      val items = downloads.values.toList()
+      val bytes = json.encodeToString(items).toByteArray()
+      val stream = file.startWrite()
+      try {
+         stream.write(bytes)
+         file.finishWrite(stream)
+      } catch (e: Exception) {
+         file.failWrite(stream)
+         Log.e(TAG, "Failed to save download store: ${e.message}")
+      }
+   }
+
+   companion object {
+      private const val TAG = "DownloadStore"
+      private const val STORE_FILENAME = "downloads.json"
+   }
+}
